refactor: modularize optical equipment base and extract calculations
Decomposed the 'God Object' file `apts/opticalequipment/abstract.py` into a modular package structure under `apts/opticalequipment/base/`.

Key changes:
- Created `apts/opticalequipment/base/equipment.py`, `intermediate.py`, and `output.py` to separate core equipment responsibilities.
- Extracted `exit_pupil` and `brightness` logic into `apts/optics/calculations.py` as pure functions, following the principle of separating data from calculation.
- Updated all internal modules and tests to import from the new modular structure.
- Standardized `super()` calls in the base classes to the modern Python 3 syntax.
- Maintained backward compatibility for the public API via `apts/opticalequipment/__init__.py`.
- Removed the now-obsolete `apts/opticalequipment/abstract.py`.

This refactor improves maintainability, reduces file-level complexity, and enhances the reusability of optical calculations.
