+++ apts/opticalequipment/base/__init__.py
@@ -1,0 +1,9 @@
+from .equipment import OpticalEquipment
+from .intermediate import IntermediateOpticalEquipment
+from .output import OutputOpticalEquipment
+
+__all__ = [
+    "OpticalEquipment",
+    "IntermediateOpticalEquipment",
+    "OutputOpticalEquipment",
+]
