+++ apts/opticalequipment/naked_eye.py
@@ -1,7 +1,7 @@
 from ..constants import GraphConstants, OpticalType
 from ..i18n import gettext_ as _
 from ..units import get_unit_registry
-from .abstract import OutputOpticalEquipment
+from .base import OutputOpticalEquipment
 
 
 class NakedEye(OutputOpticalEquipment):
