+++ apts/opticalequipment/reducer/base.py
@@ -1,4 +1,4 @@
-from ..abstract import IntermediateOpticalEquipment
+from ..base import IntermediateOpticalEquipment
 from ...constants import OpticalType
 from ...units import get_unit_registry
